requirements.txt
astroid==2.4.1
attrs==19.3.0
aws-cdk.assets==1.46.0
aws-cdk.aws-apigateway==1.41.0
aws-cdk.aws-applicationautoscaling==1.46.0
aws-cdk.aws-autoscaling==1.46.0
aws-cdk.aws-autoscaling-common==1.46.0
aws-cdk.aws-autoscaling-hooktargets==1.41.0
aws-cdk.aws-batch==1.41.0
aws-cdk.aws-certificatemanager==1.46.0
aws-cdk.aws-cloudformation==1.41.0
aws-cdk.aws-cloudfront==1.41.0
aws-cdk.aws-cloudwatch==1.46.0
aws-cdk.aws-cloudwatch-actions==1.46.0
aws-cdk.aws-codebuild==1.41.0
aws-cdk.aws-codecommit==1.41.0
aws-cdk.aws-codepipeline==1.41.0
aws-cdk.aws-cognito==1.41.0
aws-cdk.aws-ec2==1.46.0
aws-cdk.aws-ecr==1.41.0
aws-cdk.aws-ecr-assets==1.41.0
aws-cdk.aws-ecs==1.41.0
aws-cdk.aws-ecs-patterns==1.41.0
aws-cdk.aws-efs==1.41.0
aws-cdk.aws-elasticloadbalancing==1.46.0
aws-cdk.aws-elasticloadbalancingv2==1.46.0
aws-cdk.aws-events==1.46.0
aws-cdk.aws-events-targets==1.41.0
aws-cdk.aws-iam==1.46.0
aws-cdk.aws-kms==1.46.0
aws-cdk.aws-lambda==1.46.0
aws-cdk.aws-logs==1.46.0
aws-cdk.aws-route53==1.46.0
aws-cdk.aws-route53-targets==1.41.0
aws-cdk.aws-s3==1.46.0
aws-cdk.aws-s3-assets==1.46.0
aws-cdk.aws-sam==1.41.0
aws-cdk.aws-secretsmanager==1.41.0
aws-cdk.aws-servicediscovery==1.41.0
aws-cdk.aws-sns==1.46.0
aws-cdk.aws-sns-subscriptions==1.41.0
aws-cdk.aws-sqs==1.46.0
aws-cdk.aws-ssm==1.46.0
aws-cdk.aws-stepfunctions==1.41.0
aws-cdk.cdk-assets-schema==1.46.0
aws-cdk.cloud-assembly-schema==1.46.0
aws-cdk.core==1.46.0
aws-cdk.custom-resources==1.41.0
aws-cdk.cx-api==1.46.0
aws-cdk.region-info==1.46.0
cattrs==1.0.0
constructs==3.0.3
# Editable Git install with no remote (ghost-cdk==0.0.1)
-e /Users/ryandao/code/ghost-cdk/ghost_cdk
isort==4.3.21
jsii==1.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
publication==0.0.3
pylint==2.5.2
python-dateutil==2.8.1
PyYAML==5.3.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
wrapt==1.12.1
